+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Chg #50: Raise minimal PHP version to `^8.1` (@vjik)
 - Bug #56: Use file ID for attachments and embedded files (@vjik)
-- Chg #49: Adapt to Yii Mailer 6, remove `Message` class (@vjik) 
+- Chg #49, #58: Adapt to Yii Mailer 6, remove `Message` class (@vjik) 
 
 ## 3.0.1 May 24, 2024